New version of sgit - support backup extension
With the -i option one can now specify a backup extension for each file
changed. The default is still empty but now it's just like with sed:
-i.bak for instance will backup file foo as foo.bak before changing foo.

Of course for each file updated there will be a new (or edited)
additional file, the backup file. Doesn't check if the file already
exists (unless sed does that: not sure).