Typo fix -i option and moved warnings to top
With sgit version 0.0.8-1 19-04-2023 (commit
9fccaa8cd680d3ed7cd91516710a5672630b5dff) the -i option was added but
the help string showed -n, not -i.

Moved the warnings and notes about -i to right below the usage string in
README.md. The warnings are also in sgit -h both in README.md and the
script itself.

The warning about -o n (which was accidentally put in as -o -n) is now
also at the top, right under the usage string.