Make sed options an array, allowing multiple -o
This commit slightly changes the usage of -o: now one must specify - or
(for long options) --. This is an improvement and a bug fix. Originally
an oversight, it only allowed for the use of -o once (or rather each
time it was used it was overwritten). That meant that one had to quote
the options list but worse than that only one '-' was used which for
more than one option is obviously an error.

Now one can specify more than one option either by one -o (quoting the
option arg) or by using -o more than once.

New version 0.0.9-1 24-04-2023.

Significant improvements to README.md.

Updated top comments in sgit as well.