Add man page, fix exit codes, improve usage string
Added man page and updated make install rule. It will install to
/usr/local/share/man/man1 but one need not install it.

In the process of adding the man page I discovered that the usage string
was incomplete and not as clear as it could be.

Updated README.md to refer to man page and also to match the usage
string.

Something that currently is lacking from the man page is that there are
no examples. This is currently justified by the fact that the README.md
has them but this would not always be available so this will probably be
fixed in time when I have the energy, time and patience.