Improve man page - added examples and rewording
Not all examples from the README.md were added. Instead it refers to the
README.md. There are slight differences in some of the commands in that
different files are specified.

The man page clarifies what the glob '.' means. This was also done in
the README.md which will be in the next commit.