Minor fixes and updates to sgit.1
That includes in comments but also a typo fix and added the AUTHOR
section. Yes according to some it's discouraged but I don't care even if
that makes me a git to some people :-) since it is my tool and in any
case just because something is discouraged does not mean it's a bad
idea.