Fix sgit version string
The format is major.minor.patch-release DD-MM-YYYY but it had for 29
October 2023 0.0.12-1 29-04-2023. It should have been 0.0.12-1
29-10.2023 but since it's now 3 October it's (with the version updated
as it is a change even if not a functional change) 0.0.13-1 03-10-2023.

Fun fact on the number 13: the word for the fear of 13 is
triskaidekaphobia. No I didn't look this up specially: it's just that we
logophiles just have a lot of words in our mind :-)