New sgit version - add dry-run option -n
If -n is used then it will do one of two things depending on the
verbosity level: if verbosity is greater than or equal to 1 it will show
the files that would be acted on with the sed(1) commands specified if
the -n option was not used; otherwise it will only show the files that
would be acted on if -n had not been used.

There is a thought that this option will be expanded upon.

Updated man page. The README.md file has not been updated for new
examples but the usage string is updated. This will be corrected at a
later point in time.

Moved error message to AFTER the sgit usage if there is a command line
error.

New version 0.0.14 03-10-2023. Added CHANGES.md file starting with this
new version. The most recent version will be at the top of the file. I
might add previous versions to the file later on.