Check if directory is git repo first
Rather than parse args etc. before checking if the directory is a git
repository check that it's a repository first to save the user from
trouble of inputting the right options only to find out they're in the
wrong place.

If it is not a git repository it shows the usage and then the error that
it's not a git repository. Showing the usage before allows them to see
how it's used without having to be in a repository if they're not
already.