Use of -e for more than 1 sed command discouraged
Due to the fact that the -e option to sgit uses one -e for sed it can
result in an error if you don't input the commands in a certain way. For
instance although doing:

    sgit -I -o -n -e '1p;1p' README.md

will work doing something like:

    sgit -I -o -n -e '1p 1p' README.md

will not.

Therefore it is recommended that one use -e for each sed command that is
needed.