New option -t to test sed commands
Add -t option to test sed commands. It tests sed commands one at a time
so that one need not fix one command and try again only to have to fix
another problem.

With verbosity >= 1 it will show the sed command it is about to test,
first, and then run the sed command. If all is okay then it will append
'OK'; otherwise it will be whatever sed prints out.

With verbosity level < 1 it will not print out anything if there is no
error; otherwise it'll print out any errors in sed.

This does NOT test sed options as well: just sed commands.

The use of this option does not require that you specify globs. Instead
it is used in a pipe and acts on an empty string to only test the sed
command. Even if you do specify globs it will not make use of them.

If any sed command has an error it will exit non-zero; otherwise it will
exit 0. If that's not clear it means that nothing else will be done
after the testing is finished.

Major version update to 1.0.0-1 06-10-2023.