Typo fixes and fix check_man make rule
The make rule now depends on the man page existing.

There was a typo (extra space) fixed in the script.

Minor updates to comments including dedications.

Clarify in README.md that although it can work to use (by quoting) only
one -o for more than one sed option it can still be problematic and is
probably preferable to not do it.