Several enhancements and comments added
New version 1.0.0-3 22-10-2023.

When testing sed commands (sgit -t), surround commands (if showing
commands due to verbosity > 0) with 's to help make it easier to see the
commands in full. As well it has a ':' after the 'testing'.

If verbosity level is 4 or higher show the path to sed and options used.

Added examples to README.md file both for the verbosity >= 4 addition
and the change in sgit -t -v 1. Made sure that the examples that have
output match the new output.

Improve usage message some, hopefully making things a bit clearer.

With sgit -x (turn on shell tracing) don't enable right away. Instead
set a flag that will do it later so that one doesn't have tracing of the
rest of the options and globs. It also is used in the verbosity >= 4.

Comment the script a lot.

If there were other changes I don't know now.