Fix accidental glob not required with -n
I was testing this idea out but without a glob it won't do anything
anyway so it might as well require a glob as the purpose of it is to do
a dry-run on files, not a dry-run on nothing. This change was
accidentally committed, however, and this corrects it.