New release - add support for git ls-files -z
Done by -0 option.  This does not check that xargs(1) has -0. This is
the responsibility of the user. This option has not been tested well
either as I have no need for it. If a file has a space in it the debug
output will not quote it either (as always) so if you have file names
with spaces or other unsafe chars you will have to distinguish these
things for yourself.