New options and debug output improvements
The option -g git allows one to specify a git path.

The option -X xargs allows one to specify an xargs path.

The git and xargs commands are now in variables ($GIT and $XARGS
respectively).

When showing debug output filenames are double quoted (in case one has
spaces in filenames, unsafe though it is) and the same for the sed
commands (as it can be misleading otherwise especially if there are
spaces in a sed command).

Updated man page and README.md files.