Change paths to standard paths in README.md
At the suggestion of Landon Curt Noll (@lcn2) I have changed the
non-standard (MacPorts) paths in the README.md file to be standard paths
(in /usr/bin).

Obviously I used sgit itself and since I did so I have added it as an
example sgit use to show how it was done.