Fix bug when git ls-files does not find a file
In the case one were to specify a glob that does not match a file in the
repo one might see (if verbosity >= 2) something like:

    /usr/bin/sed: no input files

Instead, it now shows something like:

    debug[2]: no files found with glob: "foo"
    debug[2]: hint: do you need to use -0?

where "foo" is the glob. The double quotes are there in case of spaces
in filenames, just to help out in debugging, though in the case such a
file actually exists one would likely need to use -0 (which would run
git ls-files -z). The script hints that one might need the -0 flag but
as I note in the README.md file this is untested as I don't have any
repo with unsafe filenames.