Update Makefile, attempt test workflow again
The shellcheck rule is more elaborate and checks for .shellcheckrc
(which has now been added).

The Makefile now shows more information. There are some control
variables.

It allows one to redefine the PREFIX of the install.  The README.md
explains this, reminding that if it is used it must also be used for
uninstall.

There might have been some other changes.

If make test cannot be found (for whatever odd reason) in GitHub the
next step will be to try and run the rules that make test depends on,
individually, and if that fails, then it can be removed again.

The CodeQL yaml file has also been added to the workflows directory.