Remove codeql.yml, .gitignore and Makefile
The reason it [codeql.yml] was removed before is that GitHub does not
support shell scripts. As such I have added codeql.yml to .gitignore.
The test.yml file will have to suffice.

The Makefile rules were reordered a bit and missing targets in .PHONY
were added (and also reordered).